requirements.txt
testresources
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib